Implement put operations in terms of the gasnetc_rdma_put() routine. Gets are still done with AMs. Memsets, and barriers will always be done w/ AMs.
Removed gasnetc_pol_snd() form AMPoll().  The send queue is serviced in
more carefully located spots in the code.
Change 'src' and 'dst' arguments to gasnetc_rdma_put() from uintptr_t
to void *.  The callers all needed casts before.
When servicing the send queue, reap as many completed operations as
possible, up to a reasonable (tuneable) limit.
Add gasnetc_rdma_poll() to poll for completion of an RDMA operation.
Add tracing output when blocking or polling for an RDMA completion.
Ensure blocking for an RDMA completion will service the send queue,
even if the given operation was already completed.
Puts are now done over the gasnetc_rdma_put() routine, using the
RDMA completion counters.
Gets and memsets use the same RDMA completion counters, even though
they are implemented using AMs.
When ready to do gets using the hardware, we need only add logic for
dealing with bounce buffers for non-pinned destinations.
There is still room for improvement: cleaning up some code paths and
searching for code left-over from the reference implementation which
has become unreachable or just unneccesary.

+++ vapi-conduit/gasnet_extended.h
@@ -1,6 +1,6 @@
 /*  $Archive:: /Ti/GASNet/extended-ref/gasnet_extended.h                  $
- *     $Date: 2003/04/14 21:44:57 $
- * $Revision: 1.1.2.1 $
+ *     $Date: 2003/04/15 21:08:04 $
+ * $Revision: 1.1.2.2 $
  * Description: GASNet Extended API Header
  * Copyright 2002, Dan Bonachea <bonachea@cs.berkeley.edu>
  * Terms of use are as specified in license.txt
@@ -407,7 +407,7 @@ GASNET_INLINE_MODIFIER(gasnete_put_bulk)
 void gasnete_put_bulk (gasnet_node_t node, void* dest, void *src,
                               size_t nbytes GASNETE_THREAD_FARG) {
   gasneti_atomic_t counter = gasneti_atomic_init(0);
-  gasnetc_rdma_put(node, (uintptr_t)src, (uintptr_t)dest, nbytes, NULL, &counter);
+  gasnetc_rdma_put(node, src, dest, nbytes, NULL, &counter);
   gasnetc_rdma_wait(&counter);
 }
 GASNET_INLINE_MODIFIER(gasnete_memset)
@@ -516,7 +516,7 @@ void _gasnet_put_val(gasnet_node_t node, void *dest, gasnet_register_value_t val
     gasneti_atomic_t counter = gasneti_atomic_init(0);
     gasnet_register_value_t src = value;
     GASNETI_TRACE_PUT(PUT_VAL,node,dest,GASNETE_STARTOFBITS(&src,nbytes),nbytes);
-    gasnetc_rdma_put(node, (uintptr_t)GASNETE_STARTOFBITS(&src,nbytes), (uintptr_t)dest, nbytes, NULL, &counter);
+    gasnetc_rdma_put(node, GASNETE_STARTOFBITS(&src,nbytes), dest, nbytes, NULL, &counter);
     gasnetc_rdma_wait(&counter);
   }
 }
